requirements.txt
jinja2
livereload
beautifulsoup4
lxml
Pillow
pillow-heif
